# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Greenhouse Backend API
  description: 농장 관리 시스템 Backend API
  version: 0.0.1
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/inbound-records:
    get:
      tags:
      - inbound-record-controller
      operationId: getInboundRecords
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: inboundType
        in: query
        required: false
        schema:
          type: string
          enum:
          - FLASK_SEEDLING
          - POTTED_SEEDLING
          - PRODUCT_POT
          - SAMPLE
          - ETC
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - POTTING_PENDING
          - POTTING_IN_PROGRESS
          - PLACED
          - CANCELED
      - name: variety
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 10
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseInboundRecordResponse'
    post:
      tags:
      - inbound-record-controller
      operationId: create_5
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboundRecordCreateRequest'
        required: true
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInboundRecordResponse'
  /api/inbound-records/{inboundRecordId}/potting:
    post:
      tags:
      - inbound-record-controller
      operationId: potting
      parameters:
      - name: inboundRecordId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboundRecordPottingRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInboundRecordResponse'
      deprecated: true
  /api/inbound-records/{inboundRecordId}/potting-void:
    post:
      tags:
      - inbound-record-controller
      operationId: voidPotting
      parameters:
      - name: inboundRecordId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboundRecordPottingVoidRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInboundRecordResponse'
  /api/inbound-records/{inboundRecordId}/cancel:
    post:
      tags:
      - inbound-record-controller
      operationId: cancel
      parameters:
      - name: inboundRecordId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboundRecordCancelRequest'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInboundRecordResponse'
...
  /api/auction-lots/{id}/results:
    post:
      tags:
      - auction-tracking-controller
      operationId: addResult
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuctionLotResultRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionLotResponse'
  /api/auction-lots/{id}/confirm-return:
    post:
      tags:
      - auction-tracking-controller
      operationId: confirmReturn
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuctionLotReturnRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionLotResponse'
  /api/auction-lots/{id}/adjust-quantity:
    post:
      tags:
      - auction-tracking-controller
      operationId: adjust
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuctionLotAdjustmentRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionLotResponse'
...
  /api/inbound-records/{inboundRecordId}:
    get:
      tags:
      - inbound-record-controller
      operationId: getInboundRecord
      parameters:
      - name: inboundRecordId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInboundRecordResponse'
    patch:
      tags:
      - inbound-record-controller
      operationId: update_7
      parameters:
      - name: inboundRecordId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/InboundRecordUpdateRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseInboundRecordResponse'
  /api/auction-lots/{id}/status:
    patch:
      tags:
      - auction-tracking-controller
      operationId: changeStatus
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuctionLotStatusRequest'
        required: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionLotResponse'
...
  /api/auction-lots:
    get:
      tags:
      - auction-tracking-controller
      operationId: getLots
      parameters:
      - name: from
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: to
        in: query
        required: false
        schema:
          type: string
          format: date
      - name: market
        in: query
        required: false
        schema:
          type: string
      - name: variety
        in: query
        required: false
        schema:
          type: string
      - name: grade
        in: query
        required: false
        schema:
          type: string
      - name: status
        in: query
        required: false
        schema:
          type: string
          enum:
          - SHIPPED
          - WAITING
          - IN_PROGRESS
          - SOLD
          - PARTIALLY_SOLD
          - FAILED
          - REAUCTION_WAITING
          - RETURN_INFERRED
          - PARTIALLY_RETURNED
          - RETURNED
          - QUANTITY_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
      - name: reviewOnly
        in: query
        required: false
        schema:
          type: boolean
      - name: returnOnly
        in: query
        required: false
        schema:
          type: boolean
      - name: waitingOnly
        in: query
        required: false
        schema:
          type: boolean
      - name: keyword
        in: query
        required: false
        schema:
          type: string
      - name: page
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 0
      - name: size
        in: query
        required: false
        schema:
          type: integer
          format: int32
          default: 20
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponsePageResponseAuctionLotResponse'
  /api/auction-lots/{id}:
    get:
      tags:
      - auction-tracking-controller
      operationId: getLot
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionLotResponse'
  /api/auction-lots/{id}/timeline:
    get:
      tags:
      - auction-tracking-controller
      operationId: getTimeline
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponseAuctionLotResponse'
...
components:
  schemas:
...
    RepotResultOrchidGroupRequest:
      type: object
      properties:
        ageYear:
          type: integer
          format: int32
          minimum: 0
        bedZoneId:
          type: integer
          format: int64
        endPosition:
          type: number
          minimum: 0.0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        placementType:
          type: string
          maxLength: 50
          minLength: 0
        potSize:
          type: string
          maxLength: 50
          minLength: 0
        quantity:
          type: integer
          format: int32
          minimum: 1
        splitPlacementAllowed:
          type: boolean
        startPosition:
          type: number
          minimum: 0.0
        trayCount:
          type: integer
          format: int32
          minimum: 0
      required:
      - bedZoneId
      - endPosition
      - quantity
      - startPosition
...
    InboundNewVarietyRequest:
      type: object
      properties:
        defaultPotSize:
          type: string
          maxLength: 50
          minLength: 0
        genus:
          type: string
          maxLength: 100
          minLength: 0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        name:
          type: string
          maxLength: 150
          minLength: 0
      required:
      - genus
      - name
    InboundPlacementInput:
      type: object
      properties:
        ageYear:
          type: integer
          format: int32
          minimum: 0
        bedZoneId:
          type: integer
          format: int64
        endPosition:
          type: number
        placementType:
          type: string
          maxLength: 100
          minLength: 0
        potSize:
          type: string
          maxLength: 50
          minLength: 0
        quantity:
          type: integer
          format: int32
          minimum: 1
        startPosition:
          type: number
        trayCount:
          type: integer
          format: int32
          minimum: 0
      required:
      - bedZoneId
      - quantity
    InboundRecordCreateRequest:
      type: object
      properties:
        estimatedQuantity:
          type: integer
          format: int32
          minimum: 1
        inboundDate:
          type: string
          format: date
        inboundType:
          type: string
          enum:
          - FLASK_SEEDLING
          - POTTED_SEEDLING
          - PRODUCT_POT
          - SAMPLE
          - ETC
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        newVariety:
          $ref: '#/components/schemas/InboundNewVarietyRequest'
        placement:
          $ref: '#/components/schemas/InboundPlacementInput'
        pottingDueDate:
          type: string
          format: date
        tempLocation:
          type: string
          maxLength: 255
          minLength: 0
        varietyId:
          type: integer
          format: int64
        worker:
          type: string
          maxLength: 50
          minLength: 0
      required:
      - inboundDate
      - inboundType
    ApiResponseInboundRecordResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/InboundRecordResponse'
        message:
          type: string
    InboundOrchidGroupResponse:
      type: object
      properties:
        ageYear:
          type: integer
          format: int32
        bedZoneId:
          type: integer
          format: int64
        endPosition:
          type: number
        id:
          type: integer
          format: int64
        location:
          type: string
        placementType:
          type: string
        potSize:
          type: string
        quantity:
          type: integer
          format: int32
        startPosition:
          type: number
        status:
          type: string
        trayCount:
          type: integer
          format: int32
    InboundRecordResponse:
      type: object
      properties:
        availableActions:
          type: array
          items:
            type: string
            enum:
            - CANCEL
            - VOID_POTTING
        createdAt:
          type: string
          format: date-time
        createdOrchidGroups:
          type: array
          items:
            $ref: '#/components/schemas/InboundOrchidGroupResponse'
        editable:
          type: boolean
        estimatedQuantity:
          type: integer
          format: int32
        genus:
          type: string
        id:
          type: integer
          format: int64
        inboundDate:
          type: string
          format: date
        inboundType:
          type: string
        memo:
          type: string
        pottingDate:
          type: string
          format: date
        pottingDueDate:
          type: string
          format: date
        status:
          type: string
        tempLocation:
          type: string
        updatedAt:
          type: string
          format: date-time
        varietyId:
          type: integer
          format: int64
        varietyName:
          type: string
        worker:
          type: string
    InboundRecordPottingRequest:
      type: object
      properties:
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        pottingDate:
          type: string
          format: date
        results:
          type: array
          items:
            $ref: '#/components/schemas/RepotResultOrchidGroupRequest'
          maxItems: 100
          minItems: 0
        worker:
          type: string
          maxLength: 50
          minLength: 0
      required:
      - pottingDate
      - results
    InboundRecordPottingVoidRequest:
      type: object
      properties:
        idempotencyKey:
          type: string
          maxLength: 100
          minLength: 0
        reason:
          type: string
          maxLength: 1000
          minLength: 0
      required:
      - idempotencyKey
      - reason
    InboundRecordCancelRequest:
      type: object
      properties:
        idempotencyKey:
          type: string
          maxLength: 100
          minLength: 0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
...
    AuctionLotResultLineRequest:
      type: object
      properties:
        auctionGrade:
          type: string
          maxLength: 100
          minLength: 0
        inspectionStatus:
          type: string
          enum:
          - NORMAL
          - AUTO_MATCHED
          - CORRECTED_MATCH
          - MANUAL_REVIEW
          - MATCH_FAILED
          - QUANTITY_MISMATCH
          - RETURN_INFERRED
          - SOURCE_ERROR
        note:
          type: string
          maxLength: 1000
          minLength: 0
        quantity:
          type: integer
          format: int32
          minimum: 1
        unitPrice:
          type: integer
          format: int32
          minimum: 0
      required:
      - quantity
      - unitPrice
    AuctionLotResultRequest:
      type: object
      properties:
        attemptNo:
          type: integer
          format: int32
          minimum: 1
        attemptStatus:
          type: string
          enum:
          - SOLD
          - FAILED
          - PARTIALLY_SOLD
          - RETURN_INFERRED
        auctionDate:
          type: string
          format: date
        failedReason:
          type: string
          maxLength: 200
          minLength: 0
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        resultLines:
          type: array
          items:
            $ref: '#/components/schemas/AuctionLotResultLineRequest'
      required:
      - attemptStatus
      - auctionDate
    ApiResponseAuctionLotResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/AuctionLotResponse'
        message:
          type: string
    AuctionAttemptResponse:
      type: object
      properties:
        attemptNo:
          type: integer
          format: int32
        attemptStatus:
          type: string
          enum:
          - SOLD
          - FAILED
          - PARTIALLY_SOLD
          - RETURN_INFERRED
        auctionDate:
          type: string
          format: date
        failedReason:
          type: string
        id:
          type: integer
          format: int64
        memo:
          type: string
        resultLines:
          type: array
          items:
            $ref: '#/components/schemas/AuctionResultLineResponse'
    AuctionLotResponse:
      type: object
      properties:
        attempts:
          type: array
          items:
            $ref: '#/components/schemas/AuctionAttemptResponse'
        auctionMarket:
          type: string
        boxes:
          type: integer
          format: int32
        currentStatus:
          type: string
          enum:
          - SHIPPED
          - WAITING
          - IN_PROGRESS
          - SOLD
          - PARTIALLY_SOLD
          - FAILED
          - REAUCTION_WAITING
          - RETURN_INFERRED
          - PARTIALLY_RETURNED
          - RETURNED
          - QUANTITY_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
        failedCount:
          type: integer
          format: int32
        id:
          type: integer
          format: int64
        inspectionStatus:
          type: string
          enum:
          - NORMAL
          - AUTO_MATCHED
          - CORRECTED_MATCH
          - MANUAL_REVIEW
          - MATCH_FAILED
          - QUANTITY_MISMATCH
          - RETURN_INFERRED
          - SOURCE_ERROR
        itemName:
          type: string
        latestAuctionDate:
          type: string
          format: date
        memo:
          type: string
        returnConfirmableQuantity:
          type: integer
          format: int32
        returnConfirmedDate:
          type: string
          format: date
        returnedQuantity:
          type: integer
          format: int32
        shipmentDate:
          type: string
          format: date
        shipmentGrade:
          type: string
        shippedQuantity:
          type: integer
          format: int32
        soldQuantity:
          type: integer
          format: int32
        statusHistory:
          type: array
          items:
            $ref: '#/components/schemas/AuctionStatusHistoryResponse'
        totalAmount:
          type: integer
          format: int32
        varietyName:
          type: string
        waitingQuantity:
          type: integer
          format: int32
...
    AuctionStatusHistoryResponse:
      type: object
      properties:
        changedAt:
          type: string
          format: date-time
        id:
          type: integer
          format: int64
        memo:
          type: string
        newStatus:
          type: string
          enum:
          - SHIPPED
          - WAITING
          - IN_PROGRESS
          - SOLD
          - PARTIALLY_SOLD
          - FAILED
          - REAUCTION_WAITING
          - RETURN_INFERRED
          - PARTIALLY_RETURNED
          - RETURNED
          - QUANTITY_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
        previousStatus:
          type: string
          enum:
          - SHIPPED
          - WAITING
          - IN_PROGRESS
          - SOLD
          - PARTIALLY_SOLD
          - FAILED
          - REAUCTION_WAITING
          - RETURN_INFERRED
          - PARTIALLY_RETURNED
          - RETURNED
          - QUANTITY_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
        reason:
          type: string
        worker:
          type: string
    AuctionLotReturnRequest:
      type: object
      properties:
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        returnDate:
          type: string
          format: date
        returnedQuantity:
          type: integer
          format: int32
          minimum: 1
        worker:
          type: string
          maxLength: 100
          minLength: 0
      required:
      - returnDate
    AuctionLotAdjustmentRequest:
      type: object
      properties:
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        returnedQuantity:
          type: integer
          format: int32
          minimum: 0
        soldQuantity:
          type: integer
          format: int32
          minimum: 0
        waitingQuantity:
          type: integer
          format: int32
          minimum: 0
        worker:
          type: string
          maxLength: 100
          minLength: 0
      required:
      - returnedQuantity
      - soldQuantity
      - waitingQuantity
...
    InboundRecordUpdateRequest:
      type: object
      properties:
        estimatedQuantity:
          type: integer
          format: int32
          minimum: 1
        inboundDate:
          type: string
          format: date
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        pottingDueDate:
          type: string
          format: date
        tempLocation:
          type: string
          maxLength: 255
          minLength: 0
        worker:
          type: string
          maxLength: 50
          minLength: 0
      required:
      - inboundDate
    AuctionLotStatusRequest:
      type: object
      properties:
        memo:
          type: string
          maxLength: 1000
          minLength: 0
        reason:
          type: string
          maxLength: 200
          minLength: 0
        status:
          type: string
          enum:
          - SHIPPED
          - WAITING
          - IN_PROGRESS
          - SOLD
          - PARTIALLY_SOLD
          - FAILED
          - REAUCTION_WAITING
          - RETURN_INFERRED
          - PARTIALLY_RETURNED
          - RETURNED
          - QUANTITY_MISMATCH
          - REVIEW_REQUIRED
          - CANCELLED
        worker:
          type: string
          maxLength: 100
          minLength: 0
      required:
      - reason
      - status
...
    ApiResponsePageResponseInboundRecordResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PageResponseInboundRecordResponse'
        message:
          type: string
    PageResponseInboundRecordResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/InboundRecordResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...
    ApiResponsePageResponseAuctionLotResponse:
      type: object
      properties:
        data:
          $ref: '#/components/schemas/PageResponseAuctionLotResponse'
        message:
          type: string
    PageResponseAuctionLotResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/AuctionLotResponse'
        page:
          type: integer
          format: int32
        size:
          type: integer
          format: int32
        totalElements:
          type: integer
          format: int64
        totalPages:
          type: integer
          format: int32
...